Repurpose MaterialButtonCantReserve to materialButtonDisabled
This way we can reuse the button in a different context which makes it
more versatile.
